migrate trie tests to different file and make parts of trie.rs public.
